Handle fonts that don't have styles
This is starting to get messy, but because not all fonts have styles,
the String.split() call was causing it to break on those cases so this
adds a new "array_map" type thing which essentially loops through the
`fonts` list and replaces each line with an actual Map, which is
essentially equivalent to a php object. sorta.

Logger.debug has to go first in each of the cases. Implicit returns
means `fonts` would end up with its response instead of the map.